fix complie error
